backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# AI & Face Recognition
insightface>=0.7.3
opencv-python>=4.8.0
onnxruntime>=1.16.0
numpy>=1.24.0
pillow>=10.0.0

# Utilities
aiofiles>=23.2.0
python-dateutil>=2.8.0
